requirements.txt
fastapi>=0.135.1
uvicorn>=0.41.0
numpy>=2.4.3
onnxruntime>=1.24.3
grpcio>=1.78.0
grpcio-tools>=1.78.0
grpcio-reflection>=1.78.0
prometheus_client>=0.24.1
pydantic>=2.12.5
anyio>=4.12.1
python-multipart>=0.0.22
protobuf>=6.33.5
PyTurboJPEG==1.7.1
scipy>=1.15.0
Pillow>=11.0.0
tritonclient>=2.50.0
